refactor: main 하위 패키지로 이동
