+++ src/test/java/dealer/DealerTest.java
@@ -1,5 +1,6 @@
-package domain.card;
+package dealer;
 
+import domain.card.*;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -15,7 +16,7 @@ public class DealerTest {
                 .cards(Card.of(CardEmblem.EIGHT, CardDenomination.CLOVER), Card.of(CardEmblem.NINE, CardDenomination.SPADE))
                 .build();
 
-        Dealer dealer = Dealer.of(cardDeck);
+        domain.dealer.Dealer dealer = domain.dealer.Dealer.of(cardDeck);
         CardBundle cardBundle = dealer.handOutCard(2);
 
         Assertions.assertThat(cardBundle)
@@ -31,7 +32,7 @@ public class DealerTest {
                 .cards(Card.of(CardEmblem.EIGHT, CardDenomination.CLOVER), Card.of(CardEmblem.NINE, CardDenomination.SPADE))
                 .build();
 
-        Dealer dealer = Dealer.of(cardDeck);
+        domain.dealer.Dealer dealer = domain.dealer.Dealer.of(cardDeck);
         CardBundle cardBundle = dealer.handOutCard(1);
 
         Assertions.assertThat(cardBundle)
@@ -45,7 +46,7 @@ public class DealerTest {
                 .cards(Card.of(CardEmblem.EIGHT, CardDenomination.CLOVER))
                 .build();
 
-        Dealer dealer = Dealer.of(cardDeck);
+        domain.dealer.Dealer dealer = domain.dealer.Dealer.of(cardDeck);
 
         Assertions.assertThatThrownBy(() -> {
             CardBundle cardBundle = dealer.handOutCard(2);
